Verify deployment success for every project. Run tasks in localhost. New rocket project/role
Every role contains tasks at the end to verify that the deployment has been successful, the idea is to use k8s_info for this task, this is still WIP

Task are run in localhost instead of on the master nodes, which was not compatible with OCP 4.  The authentication with API will be done via k8s_auth ansible module (WIP)

A new project has been added to deploy rocket chat.

Signed-off-by: Jose Ignacio Jerez Rodriguez <jjerezro@redhat.com>