Latest 2.9 version of k8s module family is used against localhost, authentication is done via the k8s_auth module using user and password, subsequent api calls use the api key obtained.
The version of the OCP cluster is determined using curl to query the /version API point, which returns the kubernetes version, the minor number of the version representing the OCP version is saved in a variable for ease of use later in the playbook.

The <project>_state vars have been moved to the populate-ocp yaml file so they are defined before the role is actually called and can be used to decide if the role gets run or not.

Variables api_ca_cert and api_entrypoint are explicitely defined at the beginning of the playbook so they are next to the place were they are used.

k8s modules get a module defaults block that uses arguments common to all calls through the playbook

User and pasword to log in with the API server are provided via a file encrypted with ansible vault

At the end of the playbook the api key is revoked by loggin out.

The deploymentconfigs and statefulset are checked for correct completion at the end of the roles, using the k8s_info module.  At the moment if any deployment fails the playbook stops at that point.

Signedooff-by: tale <tale@localhost.localdomain>